requirements.txt
django
djangorestframework
requests
celery
redis